Refactor code and update data types across various components
- Changed `Integer` to `int` for ID fields in controllers, DTOs, and entities to use primitive types for efficiency.
- Updated `price` field from `Integer` to `double` in `ActivityDto` and `Activity` entity to allow for fractional pricing.
- Added a new endpoint in `ReservationController` to get available slots for an activity and streamlined the corresponding security configuration.
- Removed duplicate and relocated the endpoint for fetching available slots in `ReservationController` to a more appropriate URL structure.
- General cleanup of the codebase including alignment of DTOs and entities to use consistent data types across the application.

+++ src/main/java/dat3/adventureXP/dto/ActivityDto.java
@@ -12,11 +12,11 @@ import lombok.Setter;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class ActivityDto {
 
-     private Integer id;
+     private int id;
      private String name;
-     private Integer price;
-     private Integer ageLimit;
-     private Integer timeSlot; // 1 hour = 100, 2 hours = 200, 3 hours = 300, etc.
+     private double price;
+     private int ageLimit;
+     private int timeSlot; // 1 hour = 100, 2 hours = 200, 3 hours = 300, etc.
 
      public ActivityDto(Activity a) {
         this.id = a.getId();
